@@ -2,13 +2,9 @@
  * @flow
  */
 
-import React, {
-  Children,
-} from 'react';
-import {
-  StyleSheet,
-  View,
-} from 'react-native';
+import React, { Children } from 'react';
+import { StyleSheet, View } from 'react-native';
+import PropTypes from 'prop-types';
 import PureComponent from '../utils/PureComponent';
 import StaticContainer from 'react-static-container';
 
@@ -70,11 +66,14 @@ type Props = {
   tabBarStyle?: any,
   children: Array<React.Element<{}>>,
   navigation: ExNavigationContext,
-  onRegisterNavigatorContext: (navigatorUID: string, navigatorContext: ExNavigationTabContext) => void,
+  onRegisterNavigatorContext: (
+    navigatorUID: string,
+    navigatorContext: ExNavigationTabContext
+  ) => void,
   onUnregisterNavigatorContext: (navigatorUID: string) => void,
-  onWillChangeTab: (id: string) => bool,
+  onWillChangeTab: (id: string) => boolean,
   navigationState: Object,
-  translucent?: bool,
+  translucent?: boolean,
 };
 
 type State = {
@@ -100,12 +99,12 @@ class ExNavigationTab extends PureComponent<any, Props, State> {
   };
 
   static contextTypes = {
-    parentNavigatorUID: React.PropTypes.string,
+    parentNavigatorUID: PropTypes.string,
   };
 
   static childContextTypes = {
-    parentNavigatorUID: React.PropTypes.string,
-    navigator: React.PropTypes.instanceOf(ExNavigationTabContext),
+    parentNavigatorUID: PropTypes.string,
+    navigator: PropTypes.instanceOf(ExNavigationTabContext),
   };
 
   getChildContext() {
@@ -145,19 +144,23 @@ class ExNavigationTab extends PureComponent<any, Props, State> {
       translucent: this.props.translucent,
       style: [
         this.props.tabBarStyle,
-        this.props.tabBarColor ? {backgroundColor: this.props.tabBarColor} : {},
+        this.props.tabBarColor
+          ? { backgroundColor: this.props.tabBarColor }
+          : {},
       ],
     };
 
     const tabBar = this.props.renderTabBar(tabBarProps);
     const TabBarComponent = tabBar.type;
     // Get the tab bar's height from a static property on the class
-    const tabBarHeight =  this.props.tabBarHeight || TabBarComponent.defaultHeight || 0;
+    const tabBarHeight =
+      this.props.tabBarHeight || TabBarComponent.defaultHeight || 0;
     const isTranslucent = this.props.translucent;
 
     return (
       <View style={styles.container}>
-        <View style={{flex: 1, marginBottom: isTranslucent ? 0 : tabBarHeight}}>
+        <View
+          style={{ flex: 1, marginBottom: isTranslucent ? 0 : tabBarHeight }}>
           {this.renderTabs()}
         </View>
         {tabBar}
@@ -189,7 +192,7 @@ class ExNavigationTab extends PureComponent<any, Props, State> {
       <View
         key={tabItem.id}
         removeClippedSubviews={!isSelected}
-        style={[styles.tabContentInner, {opacity: isSelected ? 1 : 0}]}
+        style={[styles.tabContentInner, { opacity: isSelected ? 1 : 0 }]}
         pointerEvents={isSelected ? 'auto' : 'none'}>
         <StaticContainer shouldUpdate={isSelected}>
           {tabItem.element}
@@ -203,19 +206,25 @@ class ExNavigationTab extends PureComponent<any, Props, State> {
 
     this._registerNavigatorContext();
 
-    this.props.navigation.dispatch(Actions.setCurrentNavigator(
-      this.state.navigatorUID,
-      this.state.parentNavigatorUID,
-      'tab',
-      {},
-      [{
-        key: this.props.initialTab,
-      }],
-    ));
+    this.props.navigation.dispatch(
+      Actions.setCurrentNavigator(
+        this.state.navigatorUID,
+        this.state.parentNavigatorUID,
+        'tab',
+        {},
+        [
+          {
+            key: this.props.initialTab,
+          },
+        ]
+      )
+    );
   }
 
   componentWillUnmount() {
-    this.props.navigation.dispatch(Actions.removeNavigator(this.state.navigatorUID));
+    this.props.navigation.dispatch(
+      Actions.removeNavigator(this.state.navigatorUID)
+    );
     this.props.onUnregisterNavigatorContext(this.state.navigatorUID);
   }
 
@@ -226,7 +235,10 @@ class ExNavigationTab extends PureComponent<any, Props, State> {
 
     if (nextProps.navigationState !== this.props.navigationState) {
       this.setState({
-        renderedTabKeys: this._updateRenderedTabKeys(nextProps, this.state.renderedTabKeys),
+        renderedTabKeys: this._updateRenderedTabKeys(
+          nextProps,
+          this.state.renderedTabKeys
+        ),
       });
     }
   }
@@ -241,7 +253,9 @@ class ExNavigationTab extends PureComponent<any, Props, State> {
     if (prevProps.navigationState !== this.props.navigationState) {
       const navigationState = this.props.navigationState;
       const currentTabKey = navigationState.routes[navigationState.index].key;
-      const navigatorUIDForTabKey = this._getNavigatorContext().getNavigatorUIDForTabKey(currentTabKey);
+      const navigatorUIDForTabKey = this._getNavigatorContext().getNavigatorUIDForTabKey(
+        currentTabKey
+      );
       if (navigatorUIDForTabKey) {
         this.props.navigation.dispatch(
           Actions.setCurrentNavigator(navigatorUIDForTabKey)
@@ -256,7 +270,12 @@ class ExNavigationTab extends PureComponent<any, Props, State> {
     const selectedChild = navState.routes[navState.index];
 
     return [
-      ..._.uniq(_.without([...currentRenderedTabKeys, ...currentTabItems], selectedChild.key)),
+      ..._.uniq(
+        _.without(
+          [...currentRenderedTabKeys, ...currentTabItems],
+          selectedChild.key
+        )
+      ),
       selectedChild.key,
     ];
   }
@@ -265,7 +284,7 @@ class ExNavigationTab extends PureComponent<any, Props, State> {
     const tabItems = Children.map(props.children, (child, index) => {
       invariant(
         child.type === ExNavigationTabItem,
-        'All children of TabNavigation must be TabNavigationItems.',
+        'All children of TabNavigation must be TabNavigationItems.'
       );
 
       const tabItemProps = child.props;
@@ -281,8 +300,14 @@ class ExNavigationTab extends PureComponent<any, Props, State> {
         // via initialRoute
         if (child.props.initialRoute && this.props.translucent) {
           let defaultRouteConfig = child.props.defaultRouteConfig || {};
-          defaultRouteConfig = {...defaultRouteConfig, __tabBarInset: this.props.tabBarHeight};
-          tabItem.element = cloneReferencedElement(child, {...child.props, defaultRouteConfig});
+          defaultRouteConfig = {
+            ...defaultRouteConfig,
+            __tabBarInset: this.props.tabBarHeight,
+          };
+          tabItem.element = cloneReferencedElement(child, {
+            ...child.props,
+            defaultRouteConfig,
+          });
         } else {
           tabItem.element = child;
         }
@@ -341,13 +366,15 @@ class ExNavigationTab extends PureComponent<any, Props, State> {
         this.state.navigatorUID,
         this.state.parentNavigatorUID,
         this.state.id,
-        this.props.navigation,
+        this.props.navigation
       )
     );
   }
 
   _getNavigatorContext(): ExNavigationTabContext {
-    const navigatorContext: any = this.props.navigation.getNavigatorByUID(this.state.navigatorUID);
+    const navigatorContext: any = this.props.navigation.getNavigatorByUID(
+      this.state.navigatorUID
+    );
     return (navigatorContext: ExNavigationTabContext);
   }
 }
