@@ -1,16 +1,7 @@
 import React, { Component } from 'react';
-import {
-  Image,
-  StatusBar,
-  StyleSheet,
-  Text,
-  View,
-} from 'react-native';
-import {
-  TabNavigation,
-  TabNavigationItem,
-} from '@exponent/ex-navigation';
-import { Ionicons } from '@exponent/vector-icons';
+import { Image, StatusBar, StyleSheet, Text, View } from 'react-native';
+import { TabNavigation, TabNavigationItem } from '@expo/ex-navigation';
+import { Ionicons } from '@expo/vector-icons';
 
 function getColor(isSelected) {
   return isSelected ? '#fff' : '#000';
@@ -24,7 +15,7 @@ export default class TranslucentBarExample extends Component {
       tintColor: '#222',
       translucent: true,
     },
-  }
+  };
 
   render() {
     return (
@@ -40,24 +31,51 @@ export default class TranslucentBarExample extends Component {
             title="First"
             selectedStyle={styles.selectedTab}
             renderTitle={this._renderTitle}
-            renderIcon={(isSelected) => <Ionicons name="ios-boat-outline" size={24} color={getColor(isSelected)} />}>
-            <Image source={require('../assets/colorful-windows.jpg')} style={styles.image} />
+            renderIcon={isSelected => (
+              <Ionicons
+                name="ios-boat-outline"
+                size={24}
+                color={getColor(isSelected)}
+              />
+            )}>
+            <Image
+              source={require('../assets/colorful-windows.jpg')}
+              style={styles.image}
+            />
           </TabNavigationItem>
           <TabNavigationItem
             id="second"
             title="Second"
             selectedStyle={styles.selectedTab}
             renderTitle={this._renderTitle}
-            renderIcon={(isSelected) => <Ionicons name="ios-bulb-outline" size={24} color={getColor(isSelected)} />}>
-            <Image source={require('../assets/beetle.jpg')} style={styles.image} />
+            renderIcon={isSelected => (
+              <Ionicons
+                name="ios-bulb-outline"
+                size={24}
+                color={getColor(isSelected)}
+              />
+            )}>
+            <Image
+              source={require('../assets/beetle.jpg')}
+              style={styles.image}
+            />
           </TabNavigationItem>
           <TabNavigationItem
             id="third"
             title="Third"
             selectedStyle={styles.selectedTab}
             renderTitle={this._renderTitle}
-            renderIcon={(isSelected) => <Ionicons name="ios-bowtie-outline" size={24} color={getColor(isSelected)} />}>
-            <Image source={require('../assets/paintbrush.jpg')} style={styles.image} />
+            renderIcon={isSelected => (
+              <Ionicons
+                name="ios-bowtie-outline"
+                size={24}
+                color={getColor(isSelected)}
+              />
+            )}>
+            <Image
+              source={require('../assets/paintbrush.jpg')}
+              style={styles.image}
+            />
           </TabNavigationItem>
         </TabNavigation>
       </View>
@@ -66,11 +84,11 @@ export default class TranslucentBarExample extends Component {
 
   _renderTitle = (isSelected, title) => {
     return (
-      <Text style={{color: getColor(isSelected)}}>
+      <Text style={{ color: getColor(isSelected) }}>
         {title}
       </Text>
     );
-  }
+  };
 }
 
 const styles = StyleSheet.create({
@@ -89,5 +107,4 @@ const styles = StyleSheet.create({
   selectedTab: {
     backgroundColor: 'rgba(0,0,255, 0.4)',
   },
-
 });
