@@ -13,17 +13,29 @@ export { default as TabNavigationItem } from './tab/ExNavigationTabItem';
 export { default as TabBar } from './tab/ExNavigationTabBar';
 export { default as TabBadge } from './ExNavigationBadge';
 
-export { default as SlidingTabNavigation } from './sliding-tab/ExNavigationSlidingTab';
-export { default as SlidingTabNavigationItem } from './sliding-tab/ExNavigationSlidingTabItem';
+export {
+  default as SlidingTabNavigation,
+} from './sliding-tab/ExNavigationSlidingTab';
+export {
+  default as SlidingTabNavigationItem,
+} from './sliding-tab/ExNavigationSlidingTabItem';
 
 export { default as DrawerNavigation } from './drawer/ExNavigationDrawer';
-export { default as DrawerNavigationItem } from './drawer/ExNavigationDrawerItem';
+export {
+  default as DrawerNavigationItem,
+} from './drawer/ExNavigationDrawerItem';
 
 export { default as NavigationBar } from './ExNavigationBar';
 
-export { default as SharedElementOverlay } from './shared-element/ExNavigationSharedElementOverlay';
-export { default as SharedElementGroup } from './shared-element/ExNavigationSharedElementGroup';
-export { default as SharedElement } from './shared-element/ExNavigationSharedElement';
+export {
+  default as SharedElementOverlay,
+} from './shared-element/ExNavigationSharedElementOverlay';
+export {
+  default as SharedElementGroup,
+} from './shared-element/ExNavigationSharedElementGroup';
+export {
+  default as SharedElement,
+} from './shared-element/ExNavigationSharedElement';
 
 export { createNavigationEnabledStore } from './ExNavigationStore';
 
@@ -34,8 +46,13 @@ export { default as NavigationReducer } from './ExNavigationReducer';
 export * as NavigationStyles from './ExNavigationStyles';
 
 export { createRouter } from './ExNavigationRouter';
-export { withNavigation, createFocusAwareComponent } from './ExNavigationComponents';
+export {
+  withNavigation,
+  createFocusAwareComponent,
+} from './ExNavigationComponents';
 
 export { getBackButtonManager } from './ExNavigationBackButtonManager';
 
-export { default as AndroidBackButtonBehavior } from './ExNavigationAndroidBackButtonBehavior';
+export {
+  default as AndroidBackButtonBehavior,
+} from './ExNavigationAndroidBackButtonBehavior';
