@@ -1,4 +1,4 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
 import {
   Animated,
   Image,
@@ -7,29 +7,36 @@ import {
   Text,
   TouchableOpacity,
   View,
+  ViewPropTypes,
 } from 'react-native';
+import PropTypes from 'prop-types';
 import PureComponent from './utils/PureComponent';
 import { unsupportedNativeView } from './ExUnsupportedNativeView';
 import { withNavigation } from './ExNavigationComponents';
 
-let Components;
-if (global.__exponent) {
-  Components = global.__exponent.Components;
+let BlurView;
+let expoModule = global.__exponent || global.__expo;
+if (expoModule) {
+  BlurView = expoModule.BlurView
+    ? expoModule.BlurView
+    : expoModule.Components.BlurView;
 } else {
-  Components = {
-    BlurView: unsupportedNativeView('BlurView'),
-  };
+  BlurView = unsupportedNativeView('BlurView');
 }
 
 // Exponent draws under the status bar on Android, but vanilla React Native does not.
 // So we need to factor the status bar height in with Exponent but can ignore it with
 // vanilla React Native
-const STATUSBAR_HEIGHT = Platform.OS === 'ios' ? 20 : (global.__exponent ? 24 : 0);
+const STATUSBAR_HEIGHT = Platform.OS === 'ios'
+  ? 20
+  : global.__exponent ? 24 : 0;
 
 const APPBAR_HEIGHT = Platform.OS === 'ios' ? 44 : 55;
 const BACKGROUND_COLOR = Platform.OS === 'ios' ? '#EFEFF2' : '#FFF';
 const BORDER_BOTTOM_COLOR = 'rgba(0, 0, 0, .15)';
-const BORDER_BOTTOM_WIDTH = Platform.OS === 'ios' ? StyleSheet.hairlineWidth : 0;
+const BORDER_BOTTOM_WIDTH = Platform.OS === 'ios'
+  ? StyleSheet.hairlineWidth
+  : 0;
 const BACK_BUTTON_HIT_SLOP = { top: 0, bottom: 0, left: 0, right: 30 };
 
 class ExNavigationBarTitle extends PureComponent {
@@ -38,11 +45,13 @@ class ExNavigationBarTitle extends PureComponent {
 
     return (
       <View style={[titleStyles.title, style]}>
-        <Text numberOfLines={1} style={[
-          titleStyles.titleText,
-          tintColor ? {color: tintColor} : null,
-          textStyle,
-        ]}>
+        <Text
+          numberOfLines={1}
+          style={[
+            titleStyles.titleText,
+            tintColor ? { color: tintColor } : null,
+            textStyle,
+          ]}>
           {children}
         </Text>
       </View>
@@ -76,8 +85,7 @@ const titleStyles = StyleSheet.create({
   },
 });
 
-@withNavigation
-class ExNavigationBarBackButton extends PureComponent {
+@withNavigation class ExNavigationBarBackButton extends PureComponent {
   render() {
     const { tintColor } = this.props;
 
@@ -85,10 +93,9 @@ class ExNavigationBarBackButton extends PureComponent {
       <TouchableOpacity
         onPress={this._onPress}
         hitSlop={BACK_BUTTON_HIT_SLOP}
-        style={buttonStyles.buttonContainer}
-      >
+        style={buttonStyles.buttonContainer}>
         <Image
-          style={[buttonStyles.button, tintColor ? {tintColor} : null]}
+          style={[buttonStyles.button, tintColor ? { tintColor } : null]}
           source={require('./ExNavigationAssets').backIcon}
         />
       </TouchableOpacity>
@@ -103,9 +110,11 @@ class ExNavigationBarMenuButton extends PureComponent {
     const { tintColor } = this.props;
 
     return (
-      <TouchableOpacity style={buttonStyles.buttonContainer} onPress={() => this.props.navigator.toggleDrawer()}>
+      <TouchableOpacity
+        style={buttonStyles.buttonContainer}
+        onPress={() => this.props.navigator.toggleDrawer()}>
         <Image
-          style={[buttonStyles.menuButton, tintColor ? {tintColor} : null]}
+          style={[buttonStyles.menuButton, tintColor ? { tintColor } : null]}
           source={require('./ExNavigationAssets').menuIcon}
         />
       </TouchableOpacity>
@@ -172,7 +181,7 @@ export default class ExNavigationBar extends PureComponent {
     renderBackgroundComponent: PropTypes.func,
     barHeight: PropTypes.number.isRequired,
     statusBarHeight: PropTypes.number.isRequired,
-    style: View.propTypes.style,
+    style: ViewPropTypes.style,
   };
 
   constructor(props, context) {
@@ -192,7 +201,8 @@ export default class ExNavigationBar extends PureComponent {
 
     if (this.props.navigationState.index !== nextProps.navigationState.index) {
       this.setState({
-        delta: nextProps.navigationState.index - this.props.navigationState.index,
+        delta: nextProps.navigationState.index -
+          this.props.navigationState.index,
       });
     } else {
       this.setState({
@@ -224,14 +234,24 @@ export default class ExNavigationBar extends PureComponent {
     let styleFromRouteConfig = this.props.latestRoute.getBarStyle();
     let isTranslucent = !!this.props.latestRoute.getTranslucent();
     let translucentTint = this.props.latestRoute.getTranslucentTint();
-    let backgroundStyle = isTranslucent ? styles.appbarTranslucent : styles.appbarSolid;
-    let containerStyle = [styles.appbar, backgroundStyle, style, {height}, styleFromRouteConfig];
+    let backgroundStyle = isTranslucent
+      ? styles.appbarTranslucent
+      : styles.appbarSolid;
+    let containerStyle = [
+      styles.appbar,
+      backgroundStyle,
+      style,
+      { height },
+      styleFromRouteConfig,
+    ];
 
     if (this.props.overrideStyle) {
       containerStyle = [style];
     }
 
-    containerStyle.push(this.props.interpolator.forContainer(this.props, this.state.delta));
+    containerStyle.push(
+      this.props.interpolator.forContainer(this.props, this.state.delta)
+    );
 
     let leftComponents = scenesProps.map(this._renderLeft, this);
     let rightComponents = scenesProps.map(this._renderRight, this);
@@ -243,21 +263,29 @@ export default class ExNavigationBar extends PureComponent {
     });
 
     const backgroundComponents = scenesProps.map(this._renderBackground, this);
-    const wrapperStyle = [styles.wrapper, { paddingTop: APPBAR_HEIGHT + this.props.statusBarHeight }];
+    const wrapperStyle = [
+      styles.wrapper,
+      { paddingTop: APPBAR_HEIGHT + this.props.statusBarHeight },
+    ];
 
     return (
-      <View pointerEvents={this.props.visible ? 'auto' : 'none'} style={wrapperStyle}>
-        {isTranslucent && (
-          <Components.BlurView
+      <View
+        pointerEvents={this.props.visible ? 'auto' : 'none'}
+        style={wrapperStyle}>
+        {isTranslucent &&
+          <BlurView
             tint={translucentTint}
             intensity={100}
-            style={[styles.translucentUnderlay, {height}]}
-          />
-        )}
+            style={[styles.translucentUnderlay, { height }]}
+          />}
 
         <Animated.View style={containerStyle}>
           {backgroundComponents}
-          <View style={[styles.appbarInnerContainer, {top: this.props.statusBarHeight}]}>
+          <View
+            style={[
+              styles.appbarInnerContainer,
+              { top: this.props.statusBarHeight },
+            ]}>
             {titleComponents}
             {leftComponents}
             {rightComponents}
@@ -272,7 +300,7 @@ export default class ExNavigationBar extends PureComponent {
       props,
       'background',
       this.props.renderBackgroundComponent,
-      options,
+      options
     );
   }
 
@@ -281,7 +309,7 @@ export default class ExNavigationBar extends PureComponent {
       props,
       'left',
       this.props.renderLeftComponent,
-      this.props.interpolator.forLeft,
+      this.props.interpolator.forLeft
     );
   }
 
@@ -291,7 +319,7 @@ export default class ExNavigationBar extends PureComponent {
       'title',
       this.props.renderTitleComponent,
       this.props.interpolator.forCenter,
-      options,
+      options
     );
   }
 
@@ -300,27 +328,14 @@ export default class ExNavigationBar extends PureComponent {
       props,
       'right',
       this.props.renderRightComponent,
-      this.props.interpolator.forRight,
+      this.props.interpolator.forRight
     );
   }
 
-  _renderSubView(
-    props,
-    name,
-    renderer,
-    styleInterpolator,
-    options = {},
-  ) {
-    const {
-      scene,
-      navigationState,
-    } = props;
+  _renderSubView(props, name, renderer, styleInterpolator, options = {}) {
+    const { scene, navigationState } = props;
 
-    const {
-      index,
-      isStale,
-      key,
-    } = scene;
+    const { index, isStale, key } = scene;
 
     const offset = navigationState.index - index;
 
@@ -354,10 +369,7 @@ export default class ExNavigationBar extends PureComponent {
         <View
           pointerEvents={'none'}
           key={name + '_' + key}
-          style={[
-            styles[name],
-            layoutStyle,
-          ]}>
+          style={[styles[name], layoutStyle]}>
           {subView}
         </View>
       );
@@ -367,11 +379,7 @@ export default class ExNavigationBar extends PureComponent {
       <Animated.View
         pointerEvents={pointerEvents}
         key={name + '_' + key}
-        style={[
-          styles[name],
-          layoutStyle,
-          styleInterpolator(props),
-        ]}>
+        style={[styles[name], layoutStyle, styleInterpolator(props)]}>
         {subView}
       </Animated.View>
     );
