@@ -2,13 +2,11 @@
  * @flow
  */
 
-import React, { PropTypes } from 'react';
-import {
-  View,
-  StyleSheet,
-} from 'react-native';
+import React from 'react';
+import { View, StyleSheet } from 'react-native';
+import PropTypes from 'prop-types';
 
-import storeShape from 'react-redux/lib/utils/storeShape';
+import { storeShape } from 'react-redux/lib/utils/PropTypes';
 
 import Actions from './ExNavigationActions';
 import { createBackButtonManager } from './ExNavigationBackButtonManager';
@@ -46,7 +44,9 @@ export default class ExNavigationProvider extends React.Component {
     super(props);
 
     if (!props.context && !props.router) {
-      throw new Error('If no custom NavigationContext is passed into NavigationProvider, you must provide a router.');
+      throw new Error(
+        'If no custom NavigationContext is passed into NavigationProvider, you must provide a router.'
+      );
     }
 
     if (!props.context) {
