@@ -1,9 +1,5 @@
-import {
-  Animated,
-  Easing,
-  I18nManager,
-  NavigationExperimental,
-} from 'react-native';
+import { Animated, Easing, I18nManager } from 'react-native';
+import NavigationExperimental from './navigation-experimental';
 
 const {
   CardStackStyleInterpolator,
@@ -12,19 +8,35 @@ const {
 
 import type { ExNavigationStyles } from './ExNavigationTypeDefinition';
 
-const configureTimingTransition = (transitionProps, previousTransitionProps) => ({
+let useNativeDriver = false;
+
+export function setUseNativeDriverExperimental(use: boolean) {
+  useNativeDriver = use;
+}
+
+const configureTimingTransition = (
+  transitionProps,
+  previousTransitionProps
+) => ({
   timing: Animated.timing,
   easing: Easing.inOut(Easing.linear),
   duration: 150,
+  useNativeDriver,
 });
 
-const configureSpringTransition = (transitionProps, previousTransitionProps) => {
+const configureSpringTransition = (
+  transitionProps,
+  previousTransitionProps
+) => {
   let speed = 15;
   let restSpeedThreshold = 0.001;
   let restDisplacementThreshold = 0.001;
 
   // Popping should be faster than pushing
-  if (previousTransitionProps.navigationState.index >= transitionProps.navigationState.index) {
+  if (
+    previousTransitionProps.navigationState.index >=
+    transitionProps.navigationState.index
+  ) {
     speed = 40;
     restSpeedThreshold = 0.2;
     restDisplacementThreshold = 0.15;
@@ -36,22 +48,21 @@ const configureSpringTransition = (transitionProps, previousTransitionProps) => 
     speed,
     restSpeedThreshold,
     restDisplacementThreshold,
+    useNativeDriver,
   };
 };
 
 const configureNoopTransition = (transitionProps, previousTransitionProps) => ({
   timing: Animated.timing,
   duration: 1,
+  useNativeDriver,
 });
 
 /**
  * Render the initial style when the initial layout isn't measured yet.
  */
 function forInitial(props: NavigationSceneRendererProps): Object {
-  const {
-    navigationState,
-    scene,
-  } = props;
+  const { navigationState, scene } = props;
 
   const focused = navigationState.index === scene.index;
   const opacity = focused ? 1 : 0;
@@ -59,19 +70,12 @@ function forInitial(props: NavigationSceneRendererProps): Object {
   const translate = focused ? 0 : 1000000;
   return {
     opacity,
-    transform: [
-      { translateX: translate },
-      { translateY: translate },
-    ],
+    transform: [{ translateX: translate }, { translateY: translate }],
   };
 }
 
 function customForHorizontal(props: NavigationSceneRendererProps): Object {
-  const {
-    layout,
-    position,
-    scene,
-  } = props;
+  const { layout, position, scene } = props;
 
   if (!layout.isMeasured) {
     return forInitial(props);
@@ -101,10 +105,7 @@ function customForHorizontal(props: NavigationSceneRendererProps): Object {
 
   return {
     opacity,
-    transform: [
-      { translateX },
-      { translateY },
-    ],
+    transform: [{ translateX }, { translateY }],
   };
 }
 
@@ -113,12 +114,7 @@ export const SlideHorizontalIOS: ExNavigationStyles = {
   sceneAnimations: customForHorizontal,
   navigationBarAnimations: {
     forContainer: (props, delta) => {
-      const {
-        layout,
-        position,
-        scene,
-        scenes,
-      } = props;
+      const { layout, position, scene, scenes } = props;
 
       const index = scene.index;
 
@@ -129,7 +125,10 @@ export const SlideHorizontalIOS: ExNavigationStyles = {
         offset = meVisible ? offset : -offset;
       } else {
         // if we're pushing, get the previous scenes' visibility. If we're popping, get the scene ahead
-        const prevVisible = barVisibleForSceneIndex(scenes, index + (delta > 0 ? -1 : 1));
+        const prevVisible = barVisibleForSceneIndex(
+          scenes,
+          index + (delta > 0 ? -1 : 1)
+        );
         if (!prevVisible && meVisible) {
           // when showing, if a push, move from right to left, otherwise if pop, move from left to right
           offset = delta > 0 ? offset : -offset;
@@ -157,9 +156,9 @@ export const SlideHorizontalIOS: ExNavigationStyles = {
     /**
      * Crossfade the left view
      */
-    forLeft: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forLeft: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
@@ -170,9 +169,9 @@ export const SlideHorizontalIOS: ExNavigationStyles = {
     /**
      * Crossfade the title
      */
-    forCenter: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forCenter: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
@@ -191,9 +190,9 @@ export const SlideHorizontalIOS: ExNavigationStyles = {
     /**
      * Crossfade the right view
      */
-    forRight: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forRight: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
@@ -205,18 +204,12 @@ export const SlideHorizontalIOS: ExNavigationStyles = {
   gestures: CardStackPanResponder.forHorizontal,
 };
 
-
 export const SlideHorizontalFixedNav: ExNavigationStyles = {
   configureTransition: configureSpringTransition,
   sceneAnimations: customForHorizontal,
   navigationBarAnimations: {
     forContainer: (props, delta) => {
-      const {
-        layout,
-        position,
-        scene,
-        scenes,
-      } = props;
+      const { layout, position, scene, scenes } = props;
 
       const index = scene.index;
 
@@ -227,7 +220,10 @@ export const SlideHorizontalFixedNav: ExNavigationStyles = {
         offset = meVisible ? offset : -offset;
       } else {
         // if we're pushing, get the previous scenes' visibility. If we're popping, get the scene ahead
-        const prevVisible = barVisibleForSceneIndex(scenes, index + (delta > 0 ? -1 : 1));
+        const prevVisible = barVisibleForSceneIndex(
+          scenes,
+          index + (delta > 0 ? -1 : 1)
+        );
         if (!prevVisible && meVisible) {
           // when showing, if a push, move from right to left, otherwise if pop, move from left to right
           offset = delta > 0 ? offset : -offset;
@@ -255,9 +251,9 @@ export const SlideHorizontalFixedNav: ExNavigationStyles = {
     /**
      * Crossfade the left view
      */
-    forLeft: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forLeft: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
@@ -268,9 +264,9 @@ export const SlideHorizontalFixedNav: ExNavigationStyles = {
     /**
      * Crossfade the title
      */
-    forCenter: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forCenter: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
@@ -281,9 +277,9 @@ export const SlideHorizontalFixedNav: ExNavigationStyles = {
     /**
      * Crossfade the right view
      */
-    forRight: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forRight: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
@@ -292,7 +288,7 @@ export const SlideHorizontalFixedNav: ExNavigationStyles = {
       };
     },
     gestures: CardStackPanResponder.forHorizontal,
-  }
+  },
 };
 
 export const SlideHorizontal: ExNavigationStyles = {
@@ -300,12 +296,7 @@ export const SlideHorizontal: ExNavigationStyles = {
   sceneAnimations: customForHorizontal,
   navigationBarAnimations: {
     forContainer: (props, delta) => {
-      const {
-        layout,
-        position,
-        scene,
-        scenes,
-      } = props;
+      const { layout, position, scene, scenes } = props;
 
       const index = scene.index;
 
@@ -316,7 +307,10 @@ export const SlideHorizontal: ExNavigationStyles = {
         offset = meVisible ? offset : -offset;
       } else {
         // if we're pushing, get the previous scenes' visibility. If we're popping, get the scene ahead
-        const prevVisible = barVisibleForSceneIndex(scenes, index + (delta > 0 ? -1 : 1));
+        const prevVisible = barVisibleForSceneIndex(
+          scenes,
+          index + (delta > 0 ? -1 : 1)
+        );
         if (!prevVisible && meVisible) {
           // when showing, if a push, move from right to left, otherwise if pop, move from left to right
           offset = delta > 0 ? offset : -offset;
@@ -344,9 +338,9 @@ export const SlideHorizontal: ExNavigationStyles = {
     /**
      * Crossfade the left view
      */
-    forLeft: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forLeft: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index - 0.3, index, index + 0.3, index + 1],
@@ -355,7 +349,7 @@ export const SlideHorizontal: ExNavigationStyles = {
             barVisibleForSceneIndex(scenes, index) ? 0.05 : 0,
             barVisibleForSceneIndex(scenes, index) ? 1 : 0,
             barVisibleForSceneIndex(scenes, index) ? 0.3 : 0,
-            0
+            0,
           ],
         }),
       };
@@ -363,9 +357,9 @@ export const SlideHorizontal: ExNavigationStyles = {
     /**
      * Crossfade the title
      */
-    forCenter: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forCenter: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
 
       return {
         opacity: position.interpolate({
@@ -375,7 +369,7 @@ export const SlideHorizontal: ExNavigationStyles = {
             barVisibleForSceneIndex(scenes, index) ? 0.1 : 0,
             barVisibleForSceneIndex(scenes, index) ? 1 : 0,
             barVisibleForSceneIndex(scenes, index) ? 0.3 : 0,
-            0
+            0,
           ],
         }),
       };
@@ -383,9 +377,9 @@ export const SlideHorizontal: ExNavigationStyles = {
     /**
      * Crossfade the right view
      */
-    forRight: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forRight: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index - 0.3, index, index + 0.3, index + 1],
@@ -394,7 +388,7 @@ export const SlideHorizontal: ExNavigationStyles = {
             barVisibleForSceneIndex(scenes, index) ? 0.05 : 0,
             barVisibleForSceneIndex(scenes, index) ? 1 : 0,
             barVisibleForSceneIndex(scenes, index) ? 0.3 : 0,
-            0
+            0,
           ],
         }),
       };
@@ -411,11 +405,7 @@ export const SlideHorizontal: ExNavigationStyles = {
 export const FloatHorizontal = SlideHorizontal;
 
 function customForVertical(props: NavigationSceneRendererProps): Object {
-  const {
-    layout,
-    position,
-    scene,
-  } = props;
+  const { layout, position, scene } = props;
 
   if (!layout.isMeasured) {
     return forInitial(props);
@@ -432,10 +422,7 @@ function customForVertical(props: NavigationSceneRendererProps): Object {
   });
 
   return {
-    transform: [
-      { translateX },
-      { translateY },
-    ],
+    transform: [{ translateX }, { translateY }],
   };
 }
 
@@ -444,12 +431,7 @@ export const SlideVertical: ExNavigationStyles = {
   sceneAnimations: customForVertical,
   navigationBarAnimations: {
     forContainer: (props, delta) => {
-      const {
-        layout,
-        position,
-        scene,
-        scenes,
-      } = props;
+      const { layout, position, scene, scenes } = props;
 
       const index = scene.index;
 
@@ -462,7 +444,10 @@ export const SlideVertical: ExNavigationStyles = {
         offset = meVisible ? offset : 0;
         fadeOffset = meVisible ? 1 : 0;
       } else {
-        const prevVisible = barVisibleForSceneIndex(scenes, index + (delta > 0 ? -1 : 1));
+        const prevVisible = barVisibleForSceneIndex(
+          scenes,
+          index + (delta > 0 ? -1 : 1)
+        );
         if (!prevVisible && meVisible) {
           // if pushing, slide, no fade. If popping, no slide, fade
           offset = delta > 0 ? offset : 0;
@@ -500,9 +485,9 @@ export const SlideVertical: ExNavigationStyles = {
     /**
      * Crossfade the left view
      */
-    forLeft: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forLeft: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
@@ -513,9 +498,9 @@ export const SlideVertical: ExNavigationStyles = {
     /**
      * Crossfade the title
      */
-    forCenter: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forCenter: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
@@ -526,9 +511,9 @@ export const SlideVertical: ExNavigationStyles = {
     /**
      * Crossfade the right view
      */
-    forRight: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forRight: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
@@ -545,12 +530,7 @@ export const FloatVertical: ExNavigationStyles = {
   sceneAnimations: CardStackStyleInterpolator.forVertical,
   navigationBarAnimations: {
     forContainer: (props, delta) => {
-      const {
-        layout,
-        position,
-        scene,
-        scenes,
-      } = props;
+      const { layout, position, scene, scenes } = props;
 
       const index = scene.index;
 
@@ -563,7 +543,10 @@ export const FloatVertical: ExNavigationStyles = {
         offset = meVisible ? offset : 0;
         fadeOffset = meVisible ? 1 : 0;
       } else {
-        const prevVisible = barVisibleForSceneIndex(scenes, index + (delta > 0 ? -1 : 1));
+        const prevVisible = barVisibleForSceneIndex(
+          scenes,
+          index + (delta > 0 ? -1 : 1)
+        );
         if (!prevVisible && meVisible) {
           // if pushing, slide, no fade. If popping, no slide, fade
           offset = delta > 0 ? offset : 0;
@@ -601,9 +584,9 @@ export const FloatVertical: ExNavigationStyles = {
     /**
      * Crossfade the left view
      */
-    forLeft: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forLeft: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
@@ -614,9 +597,9 @@ export const FloatVertical: ExNavigationStyles = {
     /**
      * Crossfade the title
      */
-    forCenter: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forCenter: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
@@ -627,9 +610,9 @@ export const FloatVertical: ExNavigationStyles = {
     /**
      * Crossfade the right view
      */
-    forRight: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forRight: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
@@ -643,11 +626,8 @@ export const FloatVertical: ExNavigationStyles = {
 
 export const Fade: ExNavigationStyles = {
   configureTransition: configureTimingTransition,
-  sceneAnimations: (props) => {
-    const {
-      position,
-      scene,
-    } = props;
+  sceneAnimations: props => {
+    const { position, scene } = props;
 
     const index = scene.index;
     const inputRange = [index - 1, index, index + 1];
@@ -659,20 +639,12 @@ export const Fade: ExNavigationStyles = {
 
     return {
       opacity,
-      transform: [
-        { translateX: 0 },
-        { translateY: 0 },
-        { scale: 1 },
-      ],
+      transform: [{ translateX: 0 }, { translateY: 0 }, { scale: 1 }],
     };
   },
   navigationBarAnimations: {
     forContainer: (props, delta) => {
-      const {
-        position,
-        scene,
-        scenes,
-      } = props;
+      const { position, scene, scenes } = props;
 
       const index = scene.index;
 
@@ -690,9 +662,9 @@ export const Fade: ExNavigationStyles = {
     /**
      * Crossfade the left view
      */
-    forLeft: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forLeft: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
@@ -703,9 +675,9 @@ export const Fade: ExNavigationStyles = {
     /**
      * Crossfade the title
      */
-    forCenter: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forCenter: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
@@ -716,9 +688,9 @@ export const Fade: ExNavigationStyles = {
     /**
      * Crossfade the right view
      */
-    forRight: (props) => {
-      const {position, scene, scenes} = props;
-      const {index} = scene;
+    forRight: props => {
+      const { position, scene, scenes } = props;
+      const { index } = scene;
       return {
         opacity: position.interpolate({
           inputRange: [index - 1, index, index + 1],
